Add files via upload
An Online Dicionary which takes input from the user a word and finds its meaning on a locally stored JSON file and extract the meaning from it while displaying it. If the word is not found in the JSON file, it selects the most closest match of the user input in the list of keys in JSON file and displays the meaning. If the user is not satisfied with the closest match, it automatically does a google search of the input word using web automation 'selenium' package in Python and displays the output in a a chrome browser. It also provides a GUI interface for the same.