Add files via upload
It is a online web based post note application that supports a user to register and login to an account through which they can post, edit, update and delete the posts on the dashboard. This application uses django framework of python. It uses a login, register, and dashboard dynamic homepage which can be used to implement notes and the interface also allows to logout and go straight back to homepage.